requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.2.0
billiard==4.2.1
cachetools==5.5.0
celery==5.4.0
certifi==2022.12.7
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.0.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.6.7
crispy-bootstrap5==2024.10
cron-descriptor==1.4.5
cryptography==39.0.1
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.9
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==65.0.2
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.1
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
django-timezone-field==7.0
django-unfold==0.41.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
filelock==3.18.0
gunicorn==23.0.0
httplib2==0.22.0
identify==2.6.2
idna==3.4
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwcrypto==1.5.6
kombu==5.4.2
nodeenv==1.9.1
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.50
pillow==11.0.0
platformdirs==4.3.6
pre_commit==4.0.1
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
PyJWT==2.10.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.28.2
requests-oauthlib==1.3.1
rpds-py==0.21.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tablib==3.7.0
typing_extensions==4.9.0
tzdata==2024.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
whitenoise==6.7.0
